Win block moved inside openjdk block (where el and fc resides, especially *after* *static* tags)
This is supposed to fix   buggy behavior, when win build is done first, before others even start. Then, instead oc arch_notBuilt... tags, win fake tag started to appear. The change with starts may be not necessary, but even after few hours of gazing into it, I was not bale to meassure consequences of this patch