Improved versionSort method - added comment, returned deleted test
Added addDots(string) method, which adds a dot between every adjacent number
and letter so the string can be split into numeric and alphabetical strings
for better sorting ojdk7 and ojdk8 files. Alphabetical strings must be sorted
in ascending order(a first, z last) and numeric string in descending
order(9 first, 0 last) - we want the last version on the top.