support for zip files download (#42)
long live the pooling....
Right now it s imho ok. It can break, if  koji build produce together rpms and zips. This already happens, it is just coincidence that we are not yet affected.
Something like:
koji is attempted first - found correct build, but no valid rpms (but with some strange zip). Right now, it will fall to continue to another url (eg fake koji). If that build contains zip,  then after the change fake koji will not be tried.

This was nto an danger with msi and rpms -as thsoe two really should not meet. It was moreover ok with tarxz, as it is very artificial. We will see how it goes with zip.