Added container support
Iitially I was only about to remove:

            final List<String> supportedArches = new ArrayList<>(1);
            supportedArches.add("win");

But then I realised that all-arches x single arch is not at all implemented for archives
In addition, all those types of archives differs in both xml response and in file handling.

When the arch x images for url composition appered, I decided to go with the if-container switch rather. That is currently strgthforward, but if it will go wild. then isContainer switch will need to appear into job configuration/

Btw, the trick with arhitecutreed "detection" should be aplicable also for windows archives with minimal changes. If needed..ever...